Fix an N+1 preventing test using a convenient side effect to assert a reality.
The test in question asserts that after a cache miss a bunch of
associations are preloaded on a record. This doesn't actually test the
functionality in question: even if there were N queries to populate the
associations recursively down the tree, we'd still see loaded
associations after the fact. What actually matters is how many queries
were executed to prepopulate all those associations.