Remove normalized association from association cache for marshaling.
has_many associations use a CollectionProxy to act as a lazy loaded
array so that it can be used as an array or can be used to build a query.
This collection proxy will delegate all methods to either the association
or to a new scope, including respond_to?.

A normalized has_may association will have the association stored in the
association cache, since we cache the associated ids, but the association
will not be loaded.

When marshal dumping a record with normalized has_many associations, the
marshal code will call respond_to? on the collection proxy, which will load
the association to delegate respond_to? to it.  This can also corrupt
the marshal dump since it will set a new instance variable while dumping
the association, so will get interpreted as being outside the
objects instance variables.