Fix normalized has_many caching
The change to only cache attributes and associations of ActiveRecords
missed normalized (non-embedded) has_many associations. Specifically,
the cached IDs were not actually stored in memcache.

This stores cached ID arrays for normalized has_many associations
alongside the attributes and embedded associations in memcache. This
change requires incrementing the CACHE_VERSION.

As part of this change, the fast path used for the first fetch of a
record has been removed, as it hid the problem in testing.