profile: Support running a specific runner and saving to a filename.
A common workflow would be to run the benchmarks for regression testing,
then run the profile to find out why things are slower.  In this case, we
want to profile a specific runner which was most affected by a regression.
The profile identifies which methods are taking the most time, but we often
want to run stackprof with a --method option to find out which lines it
spends the most time in, so we should be able to save the stackprof dump.

Example usage:
  rake profile:run RUNNER=FetchNormalizedHitRunner FILENAME=stackprof-cpu.dump