Get the build passing on 4.2
The YAML format used by Active Record has changed as of Rails 4.2. A few
tests would not pass no matter what, as the objects being serialized
didn't implement equality comparison. I've fixed this upstream in
https://github.com/rails/rails/commit/9e25e0e1735f2ccca69679243aa8cf0885104164
and
https://github.com/rails/rails/commit/cfb371e96a2789790359ec1db779df490fd253d6.
I've also backported those fixes for our tests.

I'm actually wondering whether we need to bump the serialization
version. Is the contract that we never change the format that things are
serialized? Or is the contract that we accept anything serialized with
an older version? Rails 4.2 will accept the Rails 4.1 yaml as of
https://github.com/rails/rails/commit/438662ab32715d3b780fb54b0743cc2e19843086

That change is unreleased however, so maybe it's worth bumping it
anyway, just so we can be compatible with 4.2 as quickly as possible.

It's been brought up that we want to switch to a serialization format
that is independent of the Active Record version. I think this is a good
idea, and I'm going to explore it. However, I wanted to push up the
minimum possible change set to get us working w/ 4.2 first.

Fixes #233