Move delete failure detection to the cache fetcher.
Different cache stores return a different value from their delete methods,
so we don't try to detect whether it failed, since a falsey return value
could mean the key wasn't present or it couldn't send the command to the
server. This is less ambiguous for write, which is used by the cache
fetcher, so failure detection was moved into there.