Add support for blob replication.
An access pattern we see is when a sigle row becomes really hot. Because of the way memcached works, all the load goes to a single memcached server, which we have seen become slow.

This PR does not fix the issue, but mitigates it by replicating blobs under multiple keys. For 2 replicas, there is a (_n_ - 1) / _n_ chance they fall on different servers and the load is split, given a memcached ring of _n_ servers.

An alternative would be to do the replication right in the memcached client, as the distribution is computed so we could pick instead the _n_ + _k_ (mod _n_) th server instead of a random one to make sure they land on different machines.