Fix resetting of inverse association when lazily installing expiry hooks (#381)
ParentModelExpiration#check_association will set the :inverse_name on
the cached association options when the parent expiry hook is installed
when the inverse association isn't already loaded.  So we can't depend
on it already being set during fetching.

However, we can get this information from the active record association
reflection object. This is even better, because it is the information
used by active record to set the inverse association that this code is
trying to undo.