Remove require_if_necessary, since it is no longer necessary
I think this was added to support loading constants that hadn't been
autoloaded yet, back when we marshal dumped the record objects themselves
and before active support patched Marshal.load to do this.

This is no longer necessary, because we cache the attributes before they
are type cast to avoid making them dependent on active record's object
format or any constant names.