Help make error reported in #428 clearer
The NoMethodError on nil indicated that Hash#[] didn't find the
association. Using Hash#fetch provides a better error message
by also providing the association name.

Also, we should be getting the cached association from the
base cached_model instead of through the class of the instantiated
record. These can differ in the case of single table inheritance.
We should be able to rely on class_attribute to not have been set
in the child class, but I would prefer to avoid this unnecessary
indirection.