[Experimental] Worker based expiration strategy.
One of the things that make IdentityCache (IDC) hard to run at scale is how much it can
reflect load onthe database during high load events, most of the load occurs due to over-fetching
during misses while the application using IDC is enduring a high thruput event.

A possible solition for this is to change our cache population model. Currently IDC will try to find
an object (and it's embeded associations) in memcached, if the object is not in the cache IDC will try
and find the objects in the database. This simple mental model seems good enough for most things, however
it causes possibly lots of fetches from multiple workers during a high trhuput event. The problem can also
be seen in memcached itself, IDC uses `cas` to avoid writting stale data during high thruput events the
`cas` calls can become quite slow [I'll find some sample data to back this up, if this goes anywhere beyond
a WIP patch, not meant to be merged].

Changin our cache population model is a big refactor, so I consider we can do this in two steps.

1. Change the cache expiration model
2. Change the cache population model

This PR is/may be an prototype to see what a worker based expiration implementation looks like.

The idea is the following.

1. Make the expiration strategy configurable
2. On the worker based expiration model simply do _nothig_ to expire blobs inline
3. Implement a worker process that is supposed to be long lived and recevi binlog type events
4. Test to see how fast ruby can process incoming events, if ut turns out to be too slow to keep
   up with high trhuput events, try different concurrency models, pre-fork, threads, evented (?),
   implement the worker in native code? implement the worker in a 100% diff language and export the
   logic to go from blob -> key from ruby to the other language