fix: `load_multi_rows` breaks in earlier versions of ActiveRecord
Fixes a bug where `load_multi_rows` would break with the min-supported
gemspec due to earlier versions of ActiveSupport's `#pluck` failing to
accomodate redundant column names.

Fixed in https://github.com/rails/rails/pull/39264.

This is principally an issue when running `fetch_multi` with a
composite index including `id`. In these cases we simply exclude
the redundant `id` from the `pluck` call and reintroduce it in the
result at the same index.