Centralize Ruby Version to `.ruby-version`
The `.ruby-version` file is the ecosystem standard for defining a Ruby version. This PR adds the `.ruby-version` file, ensures a `required_ruby_version` is set, and removes all other references to Ruby in this repository, aligning it with the standard.
> [!IMPORTANT]
> Please verify the following before merging:

Verify that the changes in the PR meets the following requirements or adjust manually to make it compliant:
- [ ] `.ruby-version` file is present with the correct Ruby version defined
- [ ] A `required_ruby_version` in your gemspec is set
- [ ] There is no Ruby version present in the `dev.yml` Ruby task (before: `- ruby: x.x.x`, after: `- ruby`)
- [ ] There is no Ruby version/requirement referenced in the `Gemfile` (no lines with `ruby <some-version>`)
- [ ] A `Gemfile.lock` is built with that defined Ruby version
- [ ] There is no `TargetRubyVersion` defined in `rubocop.yml`
- [ ] The version of Rubocop installed is 1.61.0 or greater

This PR will be merged if there isn't any activity after 4 weeks.